Use GithubBackup namespace
- Adds a top-level module to namespace the library
- Moves backup functionality to GithubBackup::Backup
- Update gemspec
- Update bin
