+++ lib/github-backup/version.rb
@@ -1,3 +1,3 @@
-class GithubBackup
+module GithubBackup
   VERSION = "0.8.1"
 end
